0040-add-alarms-notifications

+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/eventRepository/eventRepositoryImpls/EventRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.realityexpander.tasky.agenda_feature.data.repositories.eventRepository.eventRepositoryImpls
 
+import com.realityexpander.tasky.R
 import com.realityexpander.tasky.agenda_feature.common.util.EventId
 import com.realityexpander.tasky.agenda_feature.data.common.convertersDTOEntityDomain.toDomain
 import com.realityexpander.tasky.agenda_feature.data.common.convertersDTOEntityDomain.toEntity
@@ -13,6 +14,7 @@ import com.realityexpander.tasky.agenda_feature.domain.IEventRepository
 import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
 import com.realityexpander.tasky.auth_feature.domain.IAuthRepository
 import com.realityexpander.tasky.core.presentation.util.UiText
+import com.realityexpander.tasky.core.util.ConnectivityObserver.InternetConnectivityObserverImpl.Companion.isInternetReachable
 import com.realityexpander.tasky.core.util.rethrowIfCancellation
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -33,6 +35,8 @@ class EventRepositoryImpl(
                 eventDao.createEvent(event.copy(isSynced = false).toEntity())  // save to local DB first
                 syncRepository.addCreatedSyncItem(event)
             }
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             val response = eventApi.createEvent(event.toEventDTOCreate())
 
@@ -70,6 +74,14 @@ class EventRepositoryImpl(
         return eventDao.getEventById(eventId)?.toDomain()
     }
 
+    override fun getEventsForRemindAtDateTimeRangeFlow(from: ZonedDateTime, to: ZonedDateTime): Flow<List<AgendaItem.Event>> {
+        return eventDao.getLocalEventsForRemindAtDateTimeRangeFlow(from, to).map { eventEntities ->
+            eventEntities.map { eventEntity ->
+                eventEntity.toDomain()
+            }
+        }
+    }
+
     // • UPDATE / UPSERT
 
     override suspend fun updateEvent(event: AgendaItem.Event, isRemoteOnly: Boolean): ResultUiText<AgendaItem.Event> {
@@ -78,6 +90,8 @@ class EventRepositoryImpl(
                 eventDao.updateEvent(event.copy(isSynced = false).toEntity())  // optimistic update
                 syncRepository.addUpdatedSyncItem(event)
             }
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             val response =
                 eventApi.updateEvent(
@@ -117,6 +131,8 @@ class EventRepositoryImpl(
         return try {
             eventDao.deleteEvent(event.toEntity())
             syncRepository.addDeletedSyncItem(event)
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             // Attempt to delete on server
             val response = eventApi.deleteEvent(event.toEventDTOUpdate())
