+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/reminderRepository/reminderRepositoryImpls/ReminderRepositoryImpl.kt
@@ -4,6 +4,7 @@ import com.realityexpander.remindery.agenda_feature.data.common.convertersDTOEnt
 import com.realityexpander.remindery.agenda_feature.data.common.convertersDTOEntityDomain.toDomain
 import com.realityexpander.remindery.agenda_feature.data.common.convertersDTOEntityDomain.toEntity
 import com.realityexpander.remindery.agenda_feature.data.repositories.reminderRepository.local.IReminderDao
+import com.realityexpander.tasky.R
 import com.realityexpander.tasky.agenda_feature.common.util.ReminderId
 import com.realityexpander.tasky.agenda_feature.data.repositories.reminderRepository.remote.reminderApi.IReminderApi
 import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository.ISyncRepository
@@ -11,6 +12,7 @@ import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.IReminderRepository
 import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
 import com.realityexpander.tasky.core.presentation.util.UiText
+import com.realityexpander.tasky.core.util.ConnectivityObserver.InternetConnectivityObserverImpl.Companion.isInternetReachable
 import com.realityexpander.tasky.core.util.rethrowIfCancellation
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -31,6 +33,8 @@ class ReminderRepositoryImpl(
                 reminderDao.createReminder(reminder.copy(isSynced = false).toEntity())
                 syncRepository.addCreatedSyncItem(reminder)
             }
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             val result = reminderApi.createReminder(reminder.toDTO())
             if(result.isSuccess) {
@@ -78,13 +82,26 @@ class ReminderRepositoryImpl(
         return reminderDao.getReminderById(reminderId)?.toDomain()
     }
 
+    override fun getRemindersForRemindAtDateTimeRangeFlow(
+        from: ZonedDateTime,
+        to: ZonedDateTime
+    ): Flow<List<AgendaItem.Reminder>> {
+        return reminderDao.getLocalRemindersForRemindAtDateTimeRangeFlow(from, to).map { reminderEntities ->
+            reminderEntities.map { reminderEntity ->
+                reminderEntity.toDomain()
+            }
+        }
+    }
+
     override suspend fun updateReminder(reminder: AgendaItem.Reminder, isRemoteOnly: Boolean): ResultUiText<Void> {
         return try {
             if(!isRemoteOnly) {
                 // save to local DB first
                 reminderDao.updateReminder(reminder.copy(isSynced = false).toEntity())
+                syncRepository.addUpdatedSyncItem(reminder)
             }
-            syncRepository.addUpdatedSyncItem(reminder)
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             val result = reminderApi.updateReminder(reminder.toDTO()) // no payload from server for this
             if(result.isSuccess) {
@@ -106,6 +123,8 @@ class ReminderRepositoryImpl(
         return try {
             reminderDao.deleteReminderById(reminder.id)
             syncRepository.addDeletedSyncItem(reminder)
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             // Attempt to delete on server
             val response = reminderApi.deleteReminder(reminder.id)
