+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/taskRepository/taskRepositoryImpls/TaskRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.realityexpander.tasky.agenda_feature.data.repositories.taskRepository.taskRepositoryImpls
 
+import com.realityexpander.tasky.R
 import com.realityexpander.tasky.agenda_feature.common.util.TaskId
 import com.realityexpander.tasky.agenda_feature.data.common.convertersDTOEntityDomain.toDTO
 import com.realityexpander.tasky.agenda_feature.data.common.convertersDTOEntityDomain.toDomain
@@ -11,6 +12,7 @@ import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.ITaskRepository
 import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
 import com.realityexpander.tasky.core.presentation.util.UiText
+import com.realityexpander.tasky.core.util.ConnectivityObserver.InternetConnectivityObserverImpl.Companion.isInternetReachable
 import com.realityexpander.tasky.core.util.rethrowIfCancellation
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -31,6 +33,8 @@ class TaskRepositoryImpl(
                 syncRepository.addCreatedSyncItem(task)
             }
 
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
+
             val result = taskApi.createTask(task.toDTO())
             if(result.isSuccess) {
                 syncRepository.removeCreatedSyncItem(task)
@@ -46,6 +50,14 @@ class TaskRepositoryImpl(
 
     // • READ
 
+    override suspend fun getTask(taskId: TaskId, isLocalOnly: Boolean): AgendaItem.Task? {
+        return taskDao.getTaskById(taskId)?.toDomain()
+    }
+
+    override suspend fun getTasksForDay(zonedDateTime: ZonedDateTime): List<AgendaItem.Task> {
+        return taskDao.getTasksForDay(zonedDateTime).map { it.toDomain() }
+    }
+
     override fun getTasksForDayFlow(zonedDateTime: ZonedDateTime): Flow<List<AgendaItem.Task>> {
         return taskDao.getTasksForDayFlow(zonedDateTime).map { taskEntities ->
             taskEntities.map { taskEntity ->
@@ -54,9 +66,15 @@ class TaskRepositoryImpl(
         }
     }
 
-    override suspend fun getTask(taskId: TaskId, isLocalOnly: Boolean): AgendaItem.Task? {
-        return taskDao.getTaskById(taskId)?.toDomain()
+    override fun getTasksForRemindAtDateTimeRangeFlow(from: ZonedDateTime, to: ZonedDateTime): Flow<List<AgendaItem.Task>> {
+        return taskDao.getTasksForRemindAtDateTimeRangeFlow(from, to).map { taskEntities ->
+            taskEntities.map { taskEntity ->
+                taskEntity.toDomain()
+            }
+        }
     }
+
+    // • UPDATE / UPSERT
 
     override suspend fun updateTask(task: AgendaItem.Task, isRemoteOnly: Boolean): ResultUiText<Void> {
         return try {
@@ -64,6 +82,8 @@ class TaskRepositoryImpl(
                 taskDao.updateTask(task.copy(isSynced = false).toEntity())  // save to local DB first
                 syncRepository.addUpdatedSyncItem(task)
             }
+
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
 
             val result = taskApi.updateTask(task.toDTO())
             if(result.isSuccess) {
@@ -76,12 +96,6 @@ class TaskRepositoryImpl(
             e.rethrowIfCancellation()
             ResultUiText.Error(UiText.Str(e.localizedMessage ?: "updateTask error"))
         }
-    }
-
-    // • UPDATE / UPSERT
-
-    override suspend fun getTasksForDay(zonedDateTime: ZonedDateTime): List<AgendaItem.Task> {
-        return taskDao.getTasksForDay(zonedDateTime).map { it.toDomain() }
     }
 
     override suspend fun upsertTaskLocally(task: AgendaItem.Task): ResultUiText<Void> {
@@ -99,28 +113,24 @@ class TaskRepositoryImpl(
 
     override suspend fun deleteTask(task: AgendaItem.Task): ResultUiText<Void> {
         return try {
-            return try {
-                taskDao.deleteTaskById(task.id)
-                syncRepository.addDeletedSyncItem(task)
+            taskDao.deleteTaskById(task.id)
+            syncRepository.addDeletedSyncItem(task)
 
-                // Attempt to delete on server
-                val result = taskApi.deleteTask(task.id)
-                if (result.isSuccess) {
-                    syncRepository.removeDeletedSyncItem(task)
-                    ResultUiText.Success()
-                } else {
-                    ResultUiText.Error(UiText.Str(result.exceptionOrNull()?.localizedMessage ?: "deleteReminder error"))
-                }
-            } catch (e: Exception) {
-                e.rethrowIfCancellation()
-                ResultUiText.Error(UiText.Str(e.message ?: "deleteReminder error"))
+            if(!isInternetReachable) return ResultUiText.Error(UiText.Res(R.string.error_no_internet))
+
+            // Attempt to delete on server
+            val result = taskApi.deleteTask(task.id)
+            if (result.isSuccess) {
+                syncRepository.removeDeletedSyncItem(task)
+                ResultUiText.Success()
+            } else {
+                ResultUiText.Error(UiText.Str(result.exceptionOrNull()?.localizedMessage ?: "deleteReminder error"))
             }
-
-
         } catch (e: Exception) {
             e.rethrowIfCancellation()
             ResultUiText.Error(UiText.Str(e.message ?: "deleteReminder error"))
         }
+
     }
 
     // • CLEAR / CLEANUP
